Task 2

+++ src/main/java/vttp/ssf/assessment/eventmanagement/models/Event.java
@@ -5,12 +5,12 @@ public class Event {
     private int eventId;
     private String eventName;
     private int eventSize;
-    private int eventDate;
+    private long eventDate;
     private int participants;
     
     public Event() {
     }
-    public Event(int eventId, String eventName, int eventSize, int eventDate, int participants) {
+    public Event(int eventId, String eventName, int eventSize, long eventDate, int participants) {
         this.eventId = eventId;
         this.eventName = eventName;
         this.eventSize = eventSize;
@@ -35,10 +35,10 @@ public class Event {
     public void setEventSize(int eventSize) {
         this.eventSize = eventSize;
     }
-    public int getEventDate() {
+    public long getEventDate() {
         return eventDate;
     }
-    public void setEventDate(int eventDate) {
+    public void setEventDate(long eventDate) {
         this.eventDate = eventDate;
     }
     public int getParticipants() {
